// app/types/Details.ts
export interface Details {
  food: string;
  tests: string;
  warning?: string;
  meds: string;
  status: string;
  personality: string;
  castration: string;
  dogInteraction: string;
  catInteraction: string;
  childrenInteraction: string;
  deworming: string;
  health: string;
  story: string;
  infoForCarer: string;
  age: string;
  bed: string;
  walk: string;
}